Extend LogCategorize from abstract CategorizeBase.
