+++ modules/core/category/src/Plugin/Action/LogCategorize.php
@@ -5,15 +5,10 @@ declare(strict_types=1);
 namespace Drupal\farm_category\Plugin\Action;
 
 use Drupal\Core\Action\Attribute\Action;
-use Drupal\Core\Action\Plugin\Action\EntityActionBase;
-use Drupal\Core\Entity\EntityTypeManagerInterface;
-use Drupal\Core\Session\AccountInterface;
 use Drupal\Core\StringTranslation\TranslatableMarkup;
-use Drupal\Core\TempStore\PrivateTempStoreFactory;
-use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
- * Redirects to a form to add categories to a log.
+ * Action that assigns categories to logs.
  */
 #[Action(
   id: 'log_categorize_action',
@@ -21,58 +16,6 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
   confirm_form_route_name: 'farm_category.log_categorize_action_form',
   type: 'log',
 )]
-class LogCategorize extends EntityActionBase {
-
-  public function __construct(
-    array $configuration,
-    $plugin_id,
-    $plugin_definition,
-    EntityTypeManagerInterface $entity_type_manager,
-    protected PrivateTempStoreFactory $tempStoreFactory,
-    protected AccountInterface $currentUser,
-  ) {
-    parent::__construct($configuration, $plugin_id, $plugin_definition, $entity_type_manager);
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public static function create(ContainerInterface $container, array $configuration, $plugin_id, $plugin_definition) {
-    // @todo Use autowiring and remove this when the parent class does.
-    // @see https://www.drupal.org/project/drupal/issues/3552110
-    return new static(
-      $configuration,
-      $plugin_id,
-      $plugin_definition,
-      $container->get('entity_type.manager'),
-      $container->get('tempstore.private'),
-      $container->get('current_user')
-    );
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function executeMultiple(array $entities) {
-    /** @var \Drupal\Core\Entity\EntityInterface[] $entities */
-    $this->tempStoreFactory->get('log_categorize_confirm')->set((string) $this->currentUser->id(), $entities);
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function execute($object = NULL) {
-    $this->executeMultiple([$object]);
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function access($object, ?AccountInterface $account = NULL, $return_as_object = FALSE) {
-    $result = $object->get('category')->access('edit', $account, TRUE)
-      ->andIf($object->access('update', $account, TRUE));
-
-    return $return_as_object ? $result : $result->isAllowed();
-  }
+class LogCategorize extends CategorizeBase {
 
 }
